Move Input component up a directory
This is so we can use it in other redux form components when necessary
