+++ src/components/Input/index.js
@@ -1,0 +1,3 @@
+import Input from './Input.js';
+
+export default Input;
