+++ src/components/auth-components/registration-form.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Field, reduxForm, focus } from 'redux-form';
 import { registerUser } from '../../actions/users';
 import { login } from '../../actions/auth';
-import Input from './input';
+import Input from '../Input';
 import {
   required,
   nonEmpty,
